+++ scripts/JSRoot.hist3d.js
@@ -1245,7 +1245,7 @@ JSROOT.define(['d3', 'painter', 'base3d', 'latex', 'hist'], (d3, jsrp, THREE, lt
 
          mesh.tooltip = function(intersect) {
             if (!Number.isInteger(intersect.faceIndex)) {
-               console.error(`faceIndex not provided, three.js version ${THREE.REVISION}, expected 127`);
+               console.error(`faceIndex not provided, three.js version ${THREE.REVISION}, expected 135`);
                return null;
             }
 
@@ -2047,7 +2047,7 @@ JSROOT.define(['d3', 'painter', 'base3d', 'latex', 'hist'], (d3, jsrp, THREE, lt
 
        line.tooltip = function(intersect) {
           if (!Number.isInteger(intersect.index)) {
-             console.error(`segment index not provided, three.js version ${THREE.REVISION}, expected 127`);
+             console.error(`segment index not provided, three.js version ${THREE.REVISION}, expected 135`);
              return null;
           }
 
@@ -2524,7 +2524,7 @@ JSROOT.define(['d3', 'painter', 'base3d', 'latex', 'hist'], (d3, jsrp, THREE, lt
 
          mesh.tooltip = function(intersect) {
             if (!Number.isInteger(intersect.index)) {
-               console.error(`intersect.index not provided, three.js version ${THREE.REVISION}, expected 127`);
+               console.error(`intersect.index not provided, three.js version ${THREE.REVISION}, expected 135`);
                return null;
             }
 
@@ -2815,7 +2815,7 @@ JSROOT.define(['d3', 'painter', 'base3d', 'latex', 'hist'], (d3, jsrp, THREE, lt
 
          combined_bins.tooltip = function(intersect) {
             if (!Number.isInteger(intersect.faceIndex)) {
-               console.error(`intersect.faceIndex not provided, three.js version ${THREE.REVISION}, expected 127`);
+               console.error(`intersect.faceIndex not provided, three.js version ${THREE.REVISION}, expected 135`);
                return null;
             }
             let indx = Math.floor(intersect.faceIndex / this.bins_faces);
@@ -3110,7 +3110,7 @@ JSROOT.define(['d3', 'painter', 'base3d', 'latex', 'hist'], (d3, jsrp, THREE, lt
      * @private */
    TGraph2DPainter.prototype.graph2DTooltip = function(intersect) {
       if (!Number.isInteger(intersect.index)) {
-         console.error(`intersect.index not provided, three.js version ${THREE.REVISION}, expected 127`);
+         console.error(`intersect.index not provided, three.js version ${THREE.REVISION}, expected 135`);
          return null;
       }
 
@@ -3420,7 +3420,7 @@ JSROOT.define(['d3', 'painter', 'base3d', 'latex', 'hist'], (d3, jsrp, THREE, lt
 
          mesh.tooltip = function(intersect) {
             if (!Number.isInteger(intersect.index)) {
-               console.error(`intersect.index not provided, three.js version ${THREE.REVISION}, expected 127`);
+               console.error(`intersect.index not provided, three.js version ${THREE.REVISION}, expected 135`);
                return null;
             }
             let indx = Math.floor(intersect.index / this.nvertex);